replace ava with tap
